linalg/wasm: move the WASM microbenches out of src and into the bench target
Three timing modules lived in src behind #[cfg(test)] + #[ignore], which no
other backend does. benches/wasm.rs already supersedes two of them — its own
header records that looping four kernels back-to-back biased the in-src
version — so those are dropped, and the activation bench moves across. The
numerical-consistency check between 16x1 and 32x1 is an assertion rather than
a timer, so it stays in src alongside the dispatch tests. The four copies of
the kernel-lookup helper in the bench collapse to one; the timing loops keep
their own warmup and repetition counts, which differ on purpose.
