+++ linalg/src/wasm/act.rs
@@ -215,62 +215,3 @@ mod test_wasm_sigmoid_relaxed {
 mod test_wasm_tanh_relaxed {
     tanh_frame_tests!(true, f32, crate::wasm::WasmTanh4Relaxed);
 }
-
-#[cfg(all(test, target_feature = "relaxed-simd"))]
-mod microbench_activations {
-    //! Microbench: WASM SIMD sigmoid/tanh vs the generic scalar fallback.
-    //! Sizes mirror typical RNN/transformer hidden dims (256, 512, 1024).
-    //!
-    //! Run with:
-    //!   RUSTFLAGS='-C target-feature=+simd128' \
-    //!     CARGO_TARGET_WASM32_WASIP1_RUNNER='wasmtime --env RUST_TEST_NOCAPTURE=1 --' \
-    //!     cargo test --release --target=wasm32-wasip1 -p tract-linalg \
-    //!     wasm::microbench_activations::microbench -- --nocapture --ignored
-    use crate::frame::element_wise::ElementWiseKer;
-    use std::time::Instant;
-
-    fn ns_per_call<K: ElementWiseKer<f32>>(buf: &mut [f32], iters: usize) -> f64 {
-        // Warmup
-        for _ in 0..50 {
-            K::run(buf, ());
-        }
-        let t0 = Instant::now();
-        for _ in 0..iters {
-            K::run(buf, ());
-        }
-        let elapsed = t0.elapsed();
-        elapsed.as_secs_f64() / iters as f64 * 1e9
-    }
-
-    fn bench(label: &str, n: usize, iters: usize) {
-        // Same input for both kernels — rebuild between to avoid post-clamp
-        // saturation mucking up the measurement.
-        let make = || (0..n).map(|i| ((i % 37) as f32 - 18.0) * 0.5).collect::<Vec<f32>>();
-
-        let mut buf = make();
-        let scalar_sig = ns_per_call::<crate::generic::sigmoid::SSigmoid4>(&mut buf, iters);
-        let mut buf = make();
-        let simd_sig = ns_per_call::<crate::wasm::WasmSigmoid4Relaxed>(&mut buf, iters);
-        let mut buf = make();
-        let scalar_tanh = ns_per_call::<crate::generic::tanh::STanh4>(&mut buf, iters);
-        let mut buf = make();
-        let simd_tanh = ns_per_call::<crate::wasm::WasmTanh4Relaxed>(&mut buf, iters);
-
-        eprintln!(
-            "{label} n={n} iters={iters}: \
-             sigmoid scalar={scalar_sig:.0} ns simd={simd_sig:.0} ns ({:.2}x); \
-             tanh scalar={scalar_tanh:.0} ns simd={simd_tanh:.0} ns ({:.2}x)",
-            scalar_sig / simd_sig,
-            scalar_tanh / simd_tanh,
-        );
-    }
-
-    #[test]
-    #[ignore]
-    fn microbench() {
-        eprintln!("=== WASM SIMD activations: scalar vs simd ===");
-        bench("hidden=256", 256, 5_000);
-        bench("hidden=512", 512, 3_000);
-        bench("hidden=1024", 1024, 2_000);
-    }
-}
